+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+- Class to locate a centered point inside a mask.
 - Notebook explaining how to fix the center of mass when it is outside the mask's boundaries.
 - Notebook explaining how to extract slices from images and masks files.
 - Helper class to show script execution summary.